+++ scanner/datastore/postgres/indexer_metadata_store.go
@@ -15,18 +15,24 @@ import (
 )
 
 // IndexerMetadataStore represents an indexer metadata datastore.
+//
+//go:generate mockgen-wrapper
 type IndexerMetadataStore interface {
 	// MigrateManifests populates the IndexerMetadataStore with manifests in ClairCore's manifest table
-	// which do not yet exist in the manifest_metadata table.
-	MigrateManifests(ctx context.Context) ([]string, error)
+	// which do not yet exist in the manifest_metadata table and sets the expiration to the given expiration.
+	//
+	// If not already, the expiration time will be converted to UTC timezone.
+	MigrateManifests(ctx context.Context, expiration time.Time) ([]string, error)
 	// StoreManifest stores the given manifest ID into the manifest_metadata table to be deleted by
 	// GCManifests after expiration passes.
 	//
 	// If not already, the expiration time will be converted to UTC timezone.
 	StoreManifest(ctx context.Context, manifestID string, expiration time.Time) error
-	// GCManifests deletes manifests from the manifest_metadata table with timestamps older than t (converted to UTC)
+	// ManifestExists returns if the given manifest exists in the table.
+	ManifestExists(ctx context.Context, manifestID string) (bool, error)
+	// GCManifests deletes manifests from the manifest_metadata table with timestamps older than expiration (converted to UTC)
 	// and returns their respective IDs.
-	GCManifests(ctx context.Context, t time.Time) ([]string, error)
+	GCManifests(ctx context.Context, expiration time.Time) ([]string, error)
 }
 
 type indexerMetadataStore struct {
@@ -37,7 +43,7 @@ type indexerMetadataStore struct {
 
 // IndexerMetadataStoreOpts defines options for creating an IndexerMetadataStore.
 type IndexerMetadataStoreOpts struct {
-	// IndexerStore represents the indexer.Store to query when MigrateManifests is called.
+	// IndexerStore represents the indexer.Store to query when MigrateManifests and GCManifests are called.
 	// If undefined, then MigrateManifests will fail.
 	IndexerStore indexer.Store
 }